refactor: 포인트 int -> long 형변환 리팩토링

+++ apps/commerce-api/src/main/java/com/loopers/application/user/UserCommand.java
@@ -13,7 +13,7 @@ public class UserCommand {
             String email,
             String birthDate,
             String gender,
-            Integer point
+            Long point
     ) {
         public UserModel toModel() {
             return new UserModel(
